requirements.txt
psycopg2==2.7.3.2
requests==2.21.0
python-dateutil==2.8.0
python-Levenshtein==0.12.0
